+++ src/main/java/ml/meajudadev/despesas/v1/model/ExpenseModel.java
@@ -3,16 +3,17 @@ package ml.meajudadev.despesas.v1.model;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.util.Calendar;
 
 public record ExpenseModel(
         @JsonProperty("id") int id,
         @JsonProperty("userId") int userId,
-        @JsonProperty("categoryId") int categoryId,
+        @JsonProperty("categoryId") Integer categoryId,
         @JsonProperty("description") String description,
         @JsonProperty("value") BigDecimal value,
         @JsonProperty("isFixed") boolean isFixed,
-        @JsonProperty("dueDate") Calendar dueDate,
+        @JsonProperty("dueDate") LocalDate dueDate,
         @JsonProperty("expenseType") char expenseType,
         @JsonProperty("expenseState") char expenseState
 ) {
